Fix report page
