+++ package.json
@@ -51,6 +51,7 @@
     "localforage": "^1.7.1",
     "mixpanel-browser": "^2.17.1",
     "moment": "^2.22.0",
+    "prop-types": "^15.6.1",
     "qrcode.react": "^0.8.0",
     "rc-progress": "^2.0.6",
     "react": "^16.3.0",
