+++ src/renderer/component/common/form-components/form-field.jsx
@@ -54,6 +54,8 @@ export class FormField extends React.PureComponent<Props> {
             />
           </div>
         );
+      } else if (type === 'textarea') {
+        input = <textarea className="form-field--textarea" type={type} id={name} {...inputProps} />;
       } else {
         input = <input type={type} id={name} {...inputProps} />;
       }
