+++ src/renderer/component/router/view.jsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import SettingsPage from 'page/settings';
 import HelpPage from 'page/help';
-import ReportPage from 'page/report.js';
+import ReportPage from 'page/report';
 import WalletPage from 'page/wallet';
-import GetCreditsPage from '../../page/getCredits';
 import SendReceivePage from 'page/sendCredits';
 import ShowPage from 'page/show';
 import PublishPage from 'page/publish';
@@ -17,6 +16,7 @@ import AuthPage from 'page/auth';
 import InvitePage from 'page/invite';
 import BackupPage from 'page/backup';
 import SubscriptionsPage from 'page/subscriptions';
+import GetCreditsPage from '../../page/getCredits';
 
 const route = (page, routesMap) => {
   const component = routesMap[page];
