+++ src/renderer/page/report/index.js
@@ -1,0 +1,7 @@
+import { connect } from 'react-redux';
+import { doShowSnackBar } from 'redux/actions/app';
+import ReportPage from './view';
+
+export default connect(null, {
+  doShowSnackBar,
+})(ReportPage);
